TC-65: Moved contact form to separate app and put the title and subtitles into the base template
