+++ app/dataprojects/templatetags/dataprojects_extras.py
@@ -1,5 +1,4 @@
 from django import template
-from django.utils.safestring import mark_safe
 from dataprojects.models import DataUseAgreement
 from django.conf import settings
 
@@ -7,20 +6,6 @@ from os.path import normpath, join
 import os
 
 register = template.Library()
-
-@register.simple_tag
-def modal_contact_form_button(text='Contact us', classes='btn btn-primary btn-md'):
-
-    return mark_safe("""
-        <button class='contact-form-button {}'>{}</button>
-    """.format(classes, text))
-
-@register.simple_tag
-def modal_contact_form_link(text='Contact us', classes=''):
-
-    return mark_safe("""
-        <a href=# class='contact-form-button {}'>{}</a>
-    """.format(classes, text))
 
 @register.filter
 def get_dua_form_file_contents(dua_form_file):
