+++ app/dataprojects/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from .views import list_data_projects, request_access, submit_request, signout, contact_form
+from .views import list_data_projects, request_access, submit_request, signout
 
 urlpatterns = (
     url(r'^$', list_data_projects),
@@ -7,5 +7,4 @@ urlpatterns = (
     url(r'^request_access/$', request_access),
     url(r'^submit_request/$', submit_request),
     url(r'^signout/$', signout),
-    url(r'^contact_form/$', contact_form, name='contact_form'),
 )
