+++ app/hypatio/urls.py
@@ -6,5 +6,6 @@ urlpatterns = [
     url(r'^dataprojects/admin/login/', page_not_found, {'exception': Exception('Admin form login disabled.')}),
     url(r'^admin/', admin.site.urls),
     url(r'^dataprojects/', include('dataprojects.urls')),
+    url(r'^contact/', include('contact.urls')),
     url(r'^', include('dataprojects.urls')),
 ]
